Rework to bring the general constants Python out to a more global scope as opposed to a manure-specific one.
