+++ RUFAS/routines/manure/gas_emissions/gas_emissions.py
@@ -1,7 +1,7 @@
 import math
 
 from RUFAS.routines.manure.constants.gas_emission_constants import GasEmissionConstants
-from RUFAS.routines.manure.constants.general_constants import GeneralConstants
+from RUFAS.routines.general_constants import GeneralConstants
 from RUFAS.routines.manure.constants.manure_constants import ManureConstants
