+++ tests/test_general_constants.py
@@ -1,0 +1,15 @@
+from pytest import approx
+
+from RUFAS.routines.general_constants import GeneralConstants
+
+
+def test_general_constants() -> None:
+    """Tests the general constants in file general_constants.py."""
+    constants = GeneralConstants
+    assert constants.GRAMS_TO_KG == approx(0.001)
+    assert constants.LITERS_TO_CUBIC_METERS == approx(0.001)
+    assert constants.KG_TO_CUBIC_METERS == approx(0.001)
+    assert constants.DAYS_PER_YEAR == 365
+    assert constants.SECONDS_PER_DAY == 86400
+    assert constants.WATER_DENSITY_KG_PER_LITER == approx(0.997)
+    assert constants.WATER_DENSITY_KG_PER_M3 == approx(0.997 * 0.001)
