+++ tests/test_manure_module/test_gas_emissions.py
@@ -3,7 +3,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 from RUFAS.routines.manure.constants.gas_emission_constants import GasEmissionConstants
-from RUFAS.routines.manure.constants.general_constants import GeneralConstants
+from RUFAS.routines.general_constants import GeneralConstants
 from RUFAS.routines.manure.constants.manure_constants import ManureConstants
 from RUFAS.routines.manure.gas_emissions.gas_emissions import GasEmissions
